Reorganize tests for consistency
- Each source directory with tests has a __tests__ subdirectory containing test files for the code in that directory
