+++ cardstack/src/components/__tests__/OptionItem.test.tsx
@@ -1,8 +1,8 @@
 import Chance from 'chance';
 import React from 'react';
 
-import { fireEvent, render } from '../test-utils';
-import { OptionItem } from '../../src/components/OptionItem';
+import { fireEvent, render } from '../../../test/test-utils';
+import { OptionItem } from '../OptionItem';
 
 jest.mock('@cardstack/components/Icon', () => ({
   Icon: jest.fn(() => null),
