+++ cardstack/src/transaction-mapping-strategies/transaction-mapping-strategy-types/__tests__/prepaid-card-payment-strategy.test.ts
@@ -1,7 +1,7 @@
 import { PREPAID_CARD_PAYMENT_MOCK } from '@cardstack/utils/__mocks__/merchant-strategies';
 import { PrepaidCardPaymentStrategy } from '@cardstack/transaction-mapping-strategies/transaction-mapping-strategy-types/prepaid-card-payment-strategy';
 
-jest.mock('../../../src/utils', () => ({ deviceUtils: { isIOS14: false } }));
+jest.mock('../../../utils', () => ({ deviceUtils: { isIOS14: false } }));
 
 jest.mock('@rainbow-me/references', () => ({
   shitcoins: 'JSON-MOCK-RETURN',
