+++ cardstack/src/utils/__tests__/formatting-utils.test.ts
